Checking if user is logged A user must be logged to access all parts of the system

+++ app/Http/Controllers/ContactController.php
@@ -15,6 +15,22 @@ class ContactController extends Controller
 
     public function create(Request $request)
     {
+
+        if (!$request->session()->has('user')) {
+            $data = Contact::all();
+            return view('create', ['contacts' => $data, 'error' => "Only logged users have access to this functionality"]);
+        }
+
+        if (strlen($request->post('name')) <= 5) {
+            $data = Contact::all();
+            return view('create', ['contacts' => $data, 'error' => "The name must be greater than 5 letters."]);
+        }
+
+        if (strlen($request->post('contact')) != 9) {
+            $data = Contact::all();
+            return view('create', ['contacts' => $data, 'error' => "The contact number must be 9 digits."]);
+        }
+
         $contact = new Contact();
         $contact->name = $request->post('name');
         $contact->contact = $request->post('contact');
@@ -24,14 +40,24 @@ class ContactController extends Controller
         return redirect('/');
     }
 
-    public function edit($id)
+    public function edit(Request $request, $id)
     {
+        if (!$request->session()->has('user')) {
+            $data = Contact::all();
+            return view('home', ['contacts' => $data, 'error' => "Only logged users have access to this functionality"]);
+        }
+
         $data = Contact::find($id);
         return view('edit', ['contact' => $data]);
     }
 
     public function update(Request $request, $id)
     {
+        if (!$request->session()->has('user')) {
+            $data = Contact::find($id);
+            return view('edit', ['contact' => $data, 'error' => "Only logged users have access to this functionality"]);
+        }
+
         $contact = Contact::find($id);
         $contact->name = $request->post('name');
         $contact->contact = $request->post('contact');
@@ -41,8 +67,13 @@ class ContactController extends Controller
         return redirect('/');
     }
 
-    public function delete($id)
+    public function delete(Request $request, $id)
     {
+        if (!$request->session()->has('user')) {
+            $data = Contact::all();
+            return view('home', ['contacts' => $data, 'error' => "Only logged users have access to this functionality"]);
+        }
+
         Contact::find($id)->delete();
 
         return redirect()->back();
